AI Images Detector/requirements.txt
torch==2.7.1
torchvision==0.22.1
timm==1.0.24
streamlit==1.54.0
streamlit-paste-button==0.1.2
Pillow==11.0.0
numpy==2.2.6
scikit-learn==1.7.1
